requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.3
django-toolbelt==0.0.1
djangorestframework==3.2.4
psycopg2==2.6.1
static3==0.6.1
wheel==0.24.0
Pillow
whitenoise==2.0.2
